hot fix, changed int to float / function

+++ mean.c
@@ -1,4 +1,4 @@
 float mean(int a, int b)
 {
-	return (a + b) / 2;
+	return (float)(a + b) / (float)2;
 }
